Navigation data specific to user
By adding an eventlistner to the Navbar.js I the system will recognize two dispatch events on pages. One for Mon and one for Bez. When he knows what to load in, he loads in the specific SideBarData.js (SideBarDataOnd.js or other). Quite fun actually. You can add dispatch events to select the right navbar data that should be loaded on the page.